Cleanup pull out UI code to its own file
